+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added @phpstan-require-implements constraints
+
 ## v0.4.3 (2024-08-21)
 * Updated Compass dependency
 * Updated Veneer dependency and Stub