+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v0.4.4 (2024-08-27)
 * Added Generic Profile implementation
 * Added @phpstan-require-implements constraints